requirements.txt
# KNN Image Classifier - Required Python Packages
# AI Lessen Opdracht: "Beter een K-nearest neighbour dan een verre vriend"

# Core scientific computing libraries
numpy>=1.21.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Image processing
Pillow>=8.3.0

# Progress bars and utilities
tqdm>=4.62.0

# Web interface
streamlit>=1.25.0
plotly>=5.15.0

# Optional: For better visualization
ipywidgets>=7.6.0

# Optional: For Jupyter notebook support (if using notebooks)
jupyter>=1.0.0
ipython>=7.25.0 